# build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ErrorResponse": {
				"properties": {
					"isSuccess": {
						"type": "boolean",
						"enum": [
							false
						],
						"nullable": false
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"isSuccess",
					"code",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SignupRequestDTO": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"address": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateStoreRequestDTO": {
				"properties": {
					"name": {
						"type": "string"
					},
					"address": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateReviewRequestDTO": {
				"properties": {
					"userId": {
						"type": "number",
						"format": "double"
					},
					"score": {
						"type": "number",
						"format": "double"
					},
					"content": {
						"type": "string"
					}
				},
				"required": [
					"userId",
					"score",
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateMissionRequestDTO": {
				"properties": {
					"content": {
						"type": "string"
					},
					"reward": {
						"type": "number",
						"format": "double"
					},
					"deadline": {
						"type": "string"
					}
				},
				"required": [
					"content",
					"reward"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ChallengeMissionRequestDTO": {
				"properties": {
					"missionId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"missionId"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "docs",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/api/v1/users/signup": {
			"post": {
				"operationId": "Signup",
				"responses": {
					"201": {
						"description": "회원가입 성공",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"400": {
						"description": "이메일 중복 또는 잘못된 입력",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"isSuccess": false,
											"code": 400,
											"message": "이미 사용 중인 이메일입니다."
										}
									}
								}
							}
						}
					}
				},
				"description": "회원가입 API\r\n신규 유저를 등록합니다. 비밀번호는 bcrypt로 해싱되어 저장됩니다.",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SignupRequestDTO"
							}
						}
					}
				}
			}
		},
		"/api/v1/regions/{regionId}/stores": {
			"post": {
				"operationId": "CreateStore",
				"responses": {
					"201": {
						"description": "가게 추가 성공",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"400": {
						"description": "존재하지 않는 지역 또는 잘못된 입력",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"isSuccess": false,
											"code": 400,
											"message": "존재하지 않는 지역입니다."
										}
									}
								}
							}
						}
					}
				},
				"description": "특정 지역에 가게 추가 API\r\nregionId 경로 변수에 해당하는 지역에 새 가게를 등록합니다.",
				"tags": [
					"Store"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "regionId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateStoreRequestDTO"
							}
						}
					}
				}
			}
		},
		"/api/v1/stores/{storeId}/reviews": {
			"post": {
				"operationId": "CreateReview",
				"responses": {
					"201": {
						"description": "리뷰 작성 성공",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"400": {
						"description": "존재하지 않는 가게/유저 또는 점수 범위 오류",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"isSuccess": false,
											"code": 400,
											"message": "존재하지 않는 가게입니다."
										}
									}
								}
							}
						}
					}
				},
				"description": "가게에 리뷰 작성 API\r\n가게/유저 존재 여부와 점수(0~5) 범위가 검증됩니다.",
				"tags": [
					"Review"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateReviewRequestDTO"
							}
						}
					}
				}
			}
		},
		"/api/v1/users/{userId}/reviews": {
			"get": {
				"operationId": "GetMyReviews",
				"responses": {
					"200": {
						"description": "리뷰 목록 조회 성공",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"400": {
						"description": "존재하지 않는 유저",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"isSuccess": false,
											"code": 400,
											"message": "존재하지 않는 유저입니다."
										}
									}
								}
							}
						}
					}
				},
				"description": "내가 작성한 리뷰 목록 조회 API",
				"tags": [
					"Review"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/stores/{storeId}/missions": {
			"post": {
				"operationId": "CreateMission",
				"responses": {
					"201": {
						"description": "미션 추가 성공",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"400": {
						"description": "존재하지 않는 가게 또는 잘못된 리워드",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"isSuccess": false,
											"code": 400,
											"message": "존재하지 않는 가게입니다."
										}
									}
								}
							}
						}
					}
				},
				"description": "가게에 미션 추가 API\r\n가게 존재 여부와 리워드(0 이상) 검증.",
				"tags": [
					"Mission"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateMissionRequestDTO"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetStoreMissions",
				"responses": {
					"200": {
						"description": "가게 미션 목록 조회 성공",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"400": {
						"description": "존재하지 않는 가게",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"isSuccess": false,
											"code": 400,
											"message": "존재하지 않는 가게입니다."
										}
									}
								}
							}
						}
					}
				},
				"description": "특정 가게의 미션 목록 조회 API",
				"tags": [
					"Mission"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/users/{userId}/missions": {
			"post": {
				"operationId": "ChallengeMission",
				"responses": {
					"201": {
						"description": "미션 도전 성공",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"400": {
						"description": "유저/미션 없음 또는 중복 도전",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"isSuccess": false,
											"code": 400,
											"message": "이미 도전 중이거나 완료한 미션입니다."
										}
									}
								}
							}
						}
					}
				},
				"description": "미션 도전하기 API\r\n같은 유저가 같은 미션을 중복 도전할 수 없습니다.",
				"tags": [
					"Mission"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChallengeMissionRequestDTO"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetMyMissions",
				"responses": {
					"200": {
						"description": "내 미션 목록 조회 성공",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"400": {
						"description": "존재하지 않는 유저",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"isSuccess": false,
											"code": 400,
											"message": "존재하지 않는 유저입니다."
										}
									}
								}
							}
						}
					}
				},
				"description": "내 미션 목록 조회 API\r\nstatus 쿼리 파라미터로 ongoing 또는 complete를 지정할 수 있습니다. 기본값은 ongoing.",
				"tags": [
					"Mission"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"ongoing",
								"complete"
							]
						}
					}
				]
			}
		},
		"/api/v1/users/{userId}/missions/{missionId}": {
			"patch": {
				"operationId": "MarkMissionComplete",
				"responses": {
					"200": {
						"description": "미션 완료 처리 성공",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"400": {
						"description": "도전 중인 미션이 아니거나 이미 완료됨",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"isSuccess": false,
											"code": 400,
											"message": "도전 중인 미션이 아닙니다."
										}
									}
								}
							}
						}
					}
				},
				"description": "미션 완료 처리 API\r\n도전 중(ongoing) 상태의 미션만 완료(complete) 상태로 변경할 수 있습니다.",
				"tags": [
					"Mission"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "missionId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}